fix: add selector which select active account setting for feature service
